[controller] Merge VeniceControllerConfig and VeniceControllerClusterConfig
