+++ services/venice-controller/src/main/java/com/linkedin/venice/controller/ZkHelixAdminClient.java
@@ -250,7 +250,7 @@ public class ZkHelixAdminClient implements HelixAdminClient {
           LeaderStandbySMD.name,
           IdealState.RebalanceMode.FULL_AUTO.toString(),
           AutoRebalanceStrategy.class.getName());
-      VeniceControllerClusterConfig config = multiClusterConfigs.getControllerConfig(clusterName);
+      VeniceControllerConfig config = multiClusterConfigs.getControllerConfig(clusterName);
       IdealState idealState = helixAdmin.getResourceIdealState(clusterName, kafkaTopic);
       // We don't set the delayed time per resource, we will use the cluster level helix config to decide
       // the delayed rebalance time
